chore(packages): Rationalized package contents in EC2 plugin
