+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/api/ec2/Health.kt
@@ -15,11 +15,14 @@
  * limitations under the License.
  *
  */
-package com.netflix.spinnaker.keel.api.ec2.cluster
+package com.netflix.spinnaker.keel.api.ec2
 
-data class Location(
-  val accountName: String,
-  val region: String,
-  val subnet: String?, // TODO: is this actually optional?
-  val availabilityZones: Set<String>
+import java.time.Duration
+
+data class Health(
+  val cooldown: Duration = Duration.ofSeconds(10),
+  val warmup: Duration = Duration.ofSeconds(600),
+  val healthCheckType: HealthCheckType = HealthCheckType.EC2,
+  val enabledMetrics: Set<Metric> = emptySet(),
+  val terminationPolicies: Set<TerminationPolicy> = setOf(TerminationPolicy.OldestInstance)
 )
