+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/api/ec2/LaunchConfiguration.kt
@@ -15,28 +15,15 @@
  * limitations under the License.
  *
  */
-package com.netflix.spinnaker.keel.api.ec2.cluster
+package com.netflix.spinnaker.keel.api.ec2
 
-import com.netflix.spinnaker.keel.api.ec2.image.ImageProvider
-
-data class LaunchConfigurationSpec(
-  val imageProvider: ImageProvider,
+data class LaunchConfiguration(
+  val imageId: String,
+  val appVersion: String,
   val instanceType: String,
   val ebsOptimized: Boolean,
   val iamRole: String,
   val keyPair: String,
   val instanceMonitoring: Boolean = false,
   val ramdiskId: String? = null
-) {
-  fun generateLaunchConfiguration(imageId: String, appVersion: String) =
-    LaunchConfiguration(
-      imageId = imageId,
-      appVersion = appVersion,
-      instanceType = instanceType,
-      ebsOptimized = ebsOptimized,
-      iamRole = iamRole,
-      keyPair = keyPair,
-      instanceMonitoring = instanceMonitoring,
-      ramdiskId = ramdiskId
-    )
-}
+)
