+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/resource/ServerGroupHandler.kt
@@ -11,13 +11,13 @@ import com.netflix.spinnaker.keel.api.ec2.HealthCheckType
 import com.netflix.spinnaker.keel.api.ec2.Metric
 import com.netflix.spinnaker.keel.api.ec2.ScalingProcess
 import com.netflix.spinnaker.keel.api.ec2.TerminationPolicy
-import com.netflix.spinnaker.keel.api.ec2.cluster.ServerGroup
-import com.netflix.spinnaker.keel.api.ec2.cluster.Dependencies
-import com.netflix.spinnaker.keel.api.ec2.cluster.Health
-import com.netflix.spinnaker.keel.api.ec2.cluster.LaunchConfiguration
-import com.netflix.spinnaker.keel.api.ec2.cluster.Location
-import com.netflix.spinnaker.keel.api.ec2.cluster.Scaling
-import com.netflix.spinnaker.keel.api.ec2.image.ArtifactAlreadyDeployedEvent
+import com.netflix.spinnaker.keel.api.ec2.ServerGroup
+import com.netflix.spinnaker.keel.api.ec2.Dependencies
+import com.netflix.spinnaker.keel.api.ec2.Health
+import com.netflix.spinnaker.keel.api.ec2.LaunchConfiguration
+import com.netflix.spinnaker.keel.api.ec2.Location
+import com.netflix.spinnaker.keel.api.ec2.Scaling
+import com.netflix.spinnaker.keel.ec2.image.ArtifactAlreadyDeployedEvent
 import com.netflix.spinnaker.keel.api.id
 import com.netflix.spinnaker.keel.api.serviceAccount
 import com.netflix.spinnaker.keel.clouddriver.CloudDriverCache
